+++ docs/mps.rst
@@ -5,10 +5,13 @@ Matrix Product States
     :maxdepth: 2
     :caption: Contents:
 	      
-.. autoclass:: tensornetwork.matrixproductstates.mps.BaseMPS
+.. autoclass:: tensornetwork.matrixproductstates.base_mps.BaseMPS
     :members:
     :noindex:
 
-.. autoclass:: tensornetwork.matrixproductstates.mps.FiniteMPS
+.. autoclass:: tensornetwork.matrixproductstates.finite_mps.FiniteMPS
+    :members:
+       
+.. autoclass:: tensornetwork.matrixproductstates.infinite_mps.InfiniteMPS
     :members:
