+++ tensornetwork/__init__.py
@@ -13,7 +13,8 @@ from tensornetwork import contractors
 from tensornetwork import config
 from typing import Text, Optional, Type, Union
 from tensornetwork.utils import load_nodes, save_nodes
-from tensornetwork.matrixproductstates.mps import FiniteMPS
+from tensornetwork.matrixproductstates.finite_mps import FiniteMPS
+from tensornetwork.matrixproductstates.infinite_mps import InfiniteMPS
 
 
 def set_default_backend(backend: Union[Text, BaseBackend]) -> None:
