+++ tensornetwork/matrixproductstates/base_mps.py
@@ -591,412 +591,3 @@ class BaseMPS:
 
   def canonicalize(self, normalize: Optional[bool] = True) -> np.number:
     raise NotImplementedError()
-
-
-class InfiniteMPS(BaseMPS):
-  """
-  An MPS class for infinite systems. 
-
-  MPS tensors are stored as a list of `Node` objects in the `InfiniteMPS.nodes`
-  attribute.
-  `InfiniteMPS` has a central site, also called orthogonality center. 
-  The position of this central site is stored in `InfiniteMPS.center_position`, 
-  and it can be be shifted using the `InfiniteMPS.position` method. 
-  `InfiniteMPS.position` uses QR and RQ methods to shift `center_position`.
-  
-  `InfiniteMPS` can be initialized either from a `list` of tensors, or
-  by calling the classmethod `InfiniteMPS.random`.
-  """
-
-  def __init__(self,
-               tensors: List[Union[BaseNode, Tensor]],
-               center_position: Optional[int] = 0,
-               connector_matrix: Optional[Union[BaseNode, Tensor]] = None,
-               backend: Optional[Text] = None) -> None:
-    """
-    Initialize a FiniteMPS.
-    Args:
-      tensors: A list of `Tensor` or `BaseNode` objects.
-      center_position: The initial position of the center site.
-      connector_matrix: A `Tensor` or `BaseNode` of rank 2 connecting
-        different unitcells. A value `None` is equivalent to an identity
-        `connector_matrix`.
-      backend: The name of the backend that should be used to perform 
-        contractions. Available backends are currently 'numpy', 'tensorflow',
-        'pytorch', 'jax'
-    """
-
-    super().__init__(
-        tensors=tensors,
-        center_position=center_position,
-        connector_matrix=connector_matrix,
-        backend=backend)
-
-  @classmethod
-  def random(cls,
-             d: List[int],
-             D: List[int],
-             dtype: Type[np.number],
-             backend: Optional[Text] = None):
-    """
-    Initialize a random `FiniteMPS`. The resulting state
-    is normalized. Its center-position is at 0.
-
-    Args:
-      d: A list of physical dimensions.
-      D: A list of bond dimensions.
-      dtype: A numpy dtype.
-      backend: An optional backend.
-    Returns:
-      `FiniteMPS`
-    """
-    #use numpy backend for tensor initialization
-    be = backend_factory.get_backend('numpy')
-    if len(D) != len(d) + 1:
-      raise ValueError('len(D) = {} is different from len(d) + 1= {}'.format(
-          len(D),
-          len(d) + 1))
-    if D[-1] != D[0]:
-      raise ValueError('D[0]={} != D[-1]={}.'.format(D[0], D[-1]))
-
-    tensors = [
-        be.randn((D[n], d[n], D[n + 1]), dtype=dtype) for n in range(len(d))
-    ]
-    return cls(tensors=tensors, center_position=0, backend=backend)
-
-  def unit_cell_transfer_operator(self, direction: Union[Text, int],
-                                  matrix: Union[BaseNode, Tensor]) -> BaseNode:
-    sites = range(len(self))
-    if direction in (-1, 'r', 'right'):
-      sites = reversed(sites)
-
-    for site in sites:
-      matrix = self.apply_transfer_operator(site, direction, matrix)
-    return matrix
-
-  def transfer_matrix_eigs(
-      self,
-      direction: Union[Text, int],
-      initial_state: Optional[Union[BaseNode, Tensor]] = None,
-      precision: Optional[float] = 1E-10,
-      num_krylov_vecs: Optional[int] = 30,
-      maxiter: Optional[int] = None):
-    """
-    Compute the dominant eigenvector of the MPS transfer matrix.
-    
-    Ars:
-      direction: 
-        * If `'1','l''left'`: return the left dominant eigenvalue
-          and eigenvector
-        * If `'-1','r''right'`: return the right dominant eigenvalue
-          and eigenvector
-      initial_state: An optional initial state.
-      num_krylov_vecs: Number of Krylov vectors to be used in `eigs`.
-      precision: The desired precision of the eigen values.
-      maxiter: The maximum number of iterations.
-    Returns:
-      `float` or `complex`: The dominant eigenvalue.
-      Node: The dominant eigenvector.
-    """
-    D = self.bond_dimensions[0]
-
-    def mv(vector):
-      result = self.unit_cell_transfer_operator(
-          direction, self.backend.reshape(vector, (D, D)))
-      return self.backend.reshape(result.tensor, (D * D,))
-
-    if not initial_state:
-      initial_state = self.backend.randn((self.bond_dimensions[0]**2,),
-                                         dtype=self.dtype)
-    else:
-      if isinstance(initial_state, BaseNode):
-        initial_state = initial_state.tensor
-      initial_state = self.backend.reshape(initial_state,
-                                           (self.bond_dimensions[0]**2,))
-
-    #note: for real dtype eta and dens are real.
-    #but scipy.linalg.eigs returns complex dtypes in any case
-    #since we know that for an MPS transfer matrix the largest
-    #eigenvalue and corresponding eigenvector are real
-    # we cast them.
-    eta, dens = self.backend.eigs(
-        A=mv,
-        initial_state=initial_state,
-        num_krylov_vecs=num_krylov_vecs,
-        numeig=1,
-        tol=precision,
-        which='LR',
-        maxiter=maxiter,
-        dtype=self.dtype)
-    result = self.backend.reshape(
-        dens[0], (self.bond_dimensions[0], self.bond_dimensions[0]))
-    return eta[0], Node(result, backend=self.backend)
-
-  def right_envs(self, sites: Sequence[int]) -> Dict:
-    raise NotImplementedError()
-
-  def left_envs(self, sites: Sequence[int]) -> Dict:
-    raise NotImplementedError()
-
-  def save(self, path: str):
-    raise NotImplementedError()
-
-
-class FiniteMPS(BaseMPS):
-  """
-  An MPS class for finite systems. 
-
-  MPS tensors are stored as a list of `Node` objects in the `FiniteMPS.nodes`
-  attribute.
-  `FiniteMPS` has a central site, also called orthogonality center. 
-  The position of this central site is stored in `FiniteMPS.center_position`, 
-  and it can be be shifted using the `FiniteMPS.position` method. 
-  `FiniteMPS.position` uses QR and RQ methods to shift `center_position`.
-  
-  `FiniteMPS` can be initialized either from a `list` of tensors, or
-  by calling the classmethod `FiniteMPS.random`.
-  
-  By default, `FiniteMPS` is initialized in *canonical* form, i.e.
-  the state is normalized, and all tensors to the left of 
-  `center_position` are left orthogonal, and all tensors 
-  to the right of `center_position` are right orthogonal. The tensor
-  at `FiniteMPS.center_position` is neither left nor right orthogonal.
-
-  Note that canonicalization can be computationally relatively 
-  costly and scales :math:`\\propto ND^3`.
-  """
-
-  def __init__(self,
-               tensors: List[Union[BaseNode, Tensor]],
-               center_position: Optional[int] = 0,
-               canonicalize: Optional[bool] = True,
-               backend: Optional[Text] = None) -> None:
-    """
-    Initialize a FiniteMPS.
-    Args:
-      tensors: A list of `Tensor` or `BaseNode` objects.
-      center_position: The initial position of the center site.
-      canonicalize: If `True` the mps is canonicalized at initialization.
-      backend: The name of the backend that should be used to perform 
-        contractions. Available backends are currently 'numpy', 'tensorflow',
-        'pytorch', 'jax'
-    """
-
-    super().__init__(
-        tensors=tensors,
-        center_position=center_position,
-        connector_matrix=None,
-        backend=backend)
-    if canonicalize:
-      if center_position == 0:
-        self.center_position = len(self) - 1
-        self.position(center_position)
-      elif center_position == len(self) - 1:
-        self.center_position = 0
-        self.position(center_position)
-      else:
-        self.center_position = 0
-        self.position(len(self) - 1)
-        self.position(center_position)
-
-  @classmethod
-  def random(cls,
-             d: List[int],
-             D: List[int],
-             dtype: Type[np.number],
-             backend: Optional[Text] = None):
-    """
-    Initialize a random `FiniteMPS`. The resulting state
-    is normalized. Its center-position is at 0.
-
-    Args:
-      d: A list of physical dimensions.
-      D: A list of bond dimensions.
-      dtype: A numpy dtype.
-      backend: An optional backend.
-    Returns:
-      `FiniteMPS`
-    """
-    #use numpy backend for tensor initialization
-    be = backend_factory.get_backend('numpy')
-    if len(D) != len(d) - 1:
-      raise ValueError('len(D) = {} is different from len(d) - 1 = {}'.format(
-          len(D),
-          len(d) - 1))
-    D = [1] + D + [1]
-    tensors = [
-        be.randn((D[n], d[n], D[n + 1]), dtype=dtype) for n in range(len(d))
-    ]
-    return cls(tensors=tensors, center_position=0, backend=backend)
-
-  def canonicalize(self, normalize: Optional[bool] = True) -> np.number:
-    """
-    Bring the MPS into canonical form according to `FiniteMPS.center_position`.
-
-    Assuming nothing about the content of the current tensors, brings the
-    tensors into canonical form with a center site at
-    `FiniteMPS.center_position`.
-
-    Args:
-      normalize: If `True`, normalize matrices when shifting.
-    Returns:
-      `Tensor`: The norm of the MPS.
-    """
-    pos = self.center_position
-    self.position(0, normalize=False)
-    self.position(len(self.nodes) - 1, normalize=False)
-    return self.position(pos, normalize=normalize)
-
-  def check_canonical(self) -> Tensor:
-    """
-    Check whether the MPS is in the expected canonical form.
-    Returns:
-      The L2 norm of the vector of local deviations.
-    """
-    deviations = []
-    for site in range(len(self.nodes)):
-      if site < self.center_position:
-        deviation = self.check_orthonormality('l', site)
-      elif site > self.center_position:
-        deviation = self.check_orthonormality('r', site)
-      else:
-        continue
-      deviations.append(deviation**2)
-    return self.backend.sqrt(sum(deviations))
-
-  def left_envs(self, sites: Sequence[int]) -> Dict:
-    """
-    Compute left reduced density matrices for site `sites`.
-    This returns a dict `left_envs` mapping sites (int) to Tensors.
-    `left_envs[site]` is the left-reduced density matrix to the left of
-    site `site`.
-
-    Args:
-      sites (list of int): A list of sites of the MPS.
-    Returns:
-      `dict` mapping `int` to `Tensor`: The left-reduced density matrices 
-        at each  site in `sites`.
-
-    """
-    n2 = max(sites)
-    sites = np.array(sites)  #enable logical indexing
-
-    #check if all elements of `sites` are within allowed range
-    if not np.all(sites <= len(self)):
-      raise ValueError('all elements of `sites` have to be <= N = {}'.format(
-          len(self)))
-    if not np.all(sites >= 0):
-      raise ValueError('all elements of `sites` have to be positive')
-
-    # left-reduced density matrices to the left of `center_position`
-    # (including center_position) are all identities
-    left_sites = sites[sites <= self.center_position]
-    left_envs = {}
-    for site in left_sites:
-      left_envs[site] = Node(
-          self.backend.eye(N=self.nodes[site].shape[0], dtype=self.dtype),
-          backend=self.backend)
-
-    # left reduced density matrices at sites > center_position
-    # have to be calculated from a network contraction
-    if n2 > self.center_position:
-      nodes = {}
-      conj_nodes = {}
-      for site in range(self.center_position, n2):
-        nodes[site] = Node(self.nodes[site], backend=self.backend)
-        conj_nodes[site] = conj(self.nodes[site])
-
-      nodes[self.center_position][0] ^ conj_nodes[self.center_position][0]
-      nodes[self.center_position][1] ^ conj_nodes[self.center_position][1]
-
-      for site in range(self.center_position + 1, n2):
-        nodes[site][0] ^ nodes[site - 1][2]
-        conj_nodes[site][0] ^ conj_nodes[site - 1][2]
-        nodes[site][1] ^ conj_nodes[site][1]
-
-      edges = {site: node[2] for site, node in nodes.items()}
-      conj_edges = {site: node[2] for site, node in conj_nodes.items()}
-
-      left_env = contract_between(nodes[self.center_position],
-                                  conj_nodes[self.center_position])
-      left_env.reorder_edges(
-          [edges[self.center_position], conj_edges[self.center_position]])
-      if self.center_position + 1 in sites:
-        left_envs[self.center_position + 1] = left_env
-      for site in range(self.center_position + 1, n2):
-        left_env = contract_between(left_env, nodes[site])
-        left_env = contract_between(left_env, conj_nodes[site])
-        if site + 1 in sites:
-          left_env.reorder_edges([edges[site], conj_edges[site]])
-          left_envs[site + 1] = left_env
-    return left_envs
-
-  def right_envs(self, sites: Sequence[int]) -> Dict:
-    """
-    Compute right reduced density matrices for site `sites.
-    This returns a dict `right_envs` mapping sites (int) to Tensors.
-    `right_envs[site]` is the right-reduced density matrix to the right of
-    site `site`.
-
-    Args:
-      sites (list of int): A list of sites of the MPS.
-    Returns:
-      `dict` mapping `int` to `Tensor`: The right-reduced density matrices 
-        at each  site in `sites`.
-    """
-
-    n1 = min(sites)
-    sites = np.array(sites)
-    #check if all elements of `sites` are within allowed range
-    if not np.all(np.array(sites) < len(self)):
-      raise ValueError('all elements of `sites` have to be < N = {}'.format(
-          len(self)))
-    if not np.all(np.array(sites) >= -1):
-      raise ValueError('all elements of `sites` have to be >= -1')
-
-    # right-reduced density matrices to the right of `center_position`
-    # (including center_position) are all identities
-    right_sites = sites[sites >= self.center_position]
-    right_envs = {}
-    for site in right_sites:
-      right_envs[site] = Node(
-          self.backend.eye(N=self.nodes[site].shape[2], dtype=self.dtype),
-          backend=self.backend)
-
-    # right reduced density matrices at sites < center_position
-    # have to be calculated from a network contraction
-    if n1 < self.center_position:
-      nodes = {}
-      conj_nodes = {}
-      for site in reversed(range(n1 + 1, self.center_position + 1)):
-        nodes[site] = Node(self.nodes[site], backend=self.backend)
-        conj_nodes[site] = conj(self.nodes[site])
-
-      nodes[self.center_position][2] ^ conj_nodes[self.center_position][2]
-      nodes[self.center_position][1] ^ conj_nodes[self.center_position][1]
-
-      for site in reversed(range(n1 + 1, self.center_position)):
-        nodes[site][2] ^ nodes[site + 1][0]
-        conj_nodes[site][2] ^ conj_nodes[site + 1][0]
-        nodes[site][1] ^ conj_nodes[site][1]
-
-      edges = {site: node[0] for site, node in nodes.items()}
-      conj_edges = {site: node[0] for site, node in conj_nodes.items()}
-
-      right_env = contract_between(nodes[self.center_position],
-                                   conj_nodes[self.center_position])
-      if self.center_position - 1 in sites:
-        right_env.reorder_edges(
-            [edges[self.center_position], conj_edges[self.center_position]])
-        right_envs[self.center_position - 1] = right_env
-      for site in reversed(range(n1 + 1, self.center_position)):
-        right_env = contract_between(right_env, nodes[site])
-        right_env = contract_between(right_env, conj_nodes[site])
-        if site - 1 in sites:
-          right_env.reorder_edges([edges[site], conj_edges[site]])
-          right_envs[site - 1] = right_env
-
-    return right_envs
-
-  def save(self, path: str):
-    raise NotImplementedError()
